+++ src/g_game.c
@@ -165,14 +165,14 @@ char            savename[MAX_PATH];
 
 gameaction_t    loadaction = ga_nothing;
 
-unsigned int    stat_gamessaved = 0;
-unsigned int    stat_mapsstarted = 0;
-unsigned int    stat_mapscompleted = 0;
-unsigned int    stat_skilllevel_imtooyoungtodie = 0;
-unsigned int    stat_skilllevel_heynottoorough = 0;
-unsigned int    stat_skilllevel_hurtmeplenty = 0;
-unsigned int    stat_skilllevel_ultraviolence = 0;
-unsigned int    stat_skilllevel_nightmare = 0;
+uint64_t        stat_gamessaved = 0;
+uint64_t        stat_mapsstarted = 0;
+uint64_t        stat_mapscompleted = 0;
+uint64_t        stat_skilllevel_imtooyoungtodie = 0;
+uint64_t        stat_skilllevel_heynottoorough = 0;
+uint64_t        stat_skilllevel_hurtmeplenty = 0;
+uint64_t        stat_skilllevel_ultraviolence = 0;
+uint64_t        stat_skilllevel_nightmare = 0;
 
 extern int      logotic;
 extern int      pagetic;
